# Optimizations for FileMatcher.GetFiles 