--- a/build/NuGetPackages/Microsoft.Build.Utilities.Core.nuspec
+++ b/build/NuGetPackages/Microsoft.Build.Utilities.Core.nuspec
@@ -21,6 +21,7 @@
         <dependency id="System.AppContext" version="4.1.0" />
         <dependency id="System.Collections" version="4.0.11" />
         <dependency id="System.Collections.Concurrent" version="4.0.12" />
+        <dependency id="System.Collections.Immutable" version="1.2.0" />
         <dependency id="System.Collections.NonGeneric" version="4.0.1" />
         <dependency id="System.Console" version="4.0.0" />
         <dependency id="System.Diagnostics.Debug" version="4.0.11" />
@@ -47,6 +48,7 @@
         <dependency id="System.Text.RegularExpressions" version="4.1.0" />
         <dependency id="System.Threading" version="4.0.11" />
         <dependency id="System.Threading.Tasks" version="4.0.11" />
+        <dependency id="System.Threading.Tasks.Parallel" version="4.0.1" />
         <dependency id="System.Threading.Thread" version="4.0.0" />
         <dependency id="System.Threading.Timer" version="4.0.1" />
         <dependency id="System.Xml.ReaderWriter" version="4.0.11" />
@@ -54,6 +56,7 @@
       </group>
       <group targetFramework=".NETFramework4.6">
         <dependency id="Microsoft.Build.Framework" version="[$version$]" />
+        <dependency id="System.Collections.Immutable" version="1.3.1" />
         <dependency id="System.Runtime.InteropServices.RuntimeInformation" version="4.3.0" />
       </group>
     </dependencies>