--- a/src/Build.UnitTests/project.json
+++ b/src/Build.UnitTests/project.json
@@ -10,6 +10,7 @@
   "frameworks": {
     "net46": {
       "Microsoft.Net.Compilers": "2.3.1",
+      "System.Collections.Immutable": "1.3.1",
       "System.Threading.Tasks.Dataflow": "4.5.24.0"
     },
     "netstandard1.3": {
@@ -17,6 +18,7 @@
         "System.Net.NetworkInformation": "4.1.0",
         "Microsoft.NETCore": "5.0.1",
         "Microsoft.NETCore.Portable.Compatibility": "1.0.1",
+        "System.Collections.Immutable": "1.2.0",
         "System.Collections.NonGeneric": "4.0.1",
         "System.Console": "4.0.0",
         "System.Diagnostics.Contracts": "4.0.1",