--- a/src/Build/project.json
+++ b/src/Build/project.json
@@ -32,6 +32,7 @@
         "System.Runtime.Serialization.Primitives": "4.1.1",
         "System.Text.Encoding.CodePages" : "4.0.1",
         "System.Threading.Tasks.Dataflow": "4.6.0.0",
+        "System.Threading.Tasks.Parallel": "4.0.1",
         "System.Threading.Thread": "4.0.0",
         "System.Threading.ThreadPool": "4.0.10",
         "System.Xml.XmlDocument": "4.0.1",