--- a/src/Tasks.UnitTests/project.json
+++ b/src/Tasks.UnitTests/project.json
@@ -12,6 +12,7 @@
     "net46": {
       "dependencies": {
         "xunit.runner.visualstudio": "2.1.0",
+        "System.Collections.Immutable": "1.3.1",
         "System.Runtime.InteropServices.RuntimeInformation": "4.3.0"
       }
     },
@@ -20,6 +21,7 @@
         "System.Net.NetworkInformation": "4.1.0",
         "Microsoft.NETCore": "5.0.1",
         "Microsoft.NETCore.Portable.Compatibility": "1.0.1",
+        "System.Collections.Immutable": "1.2.0",
         "System.Collections.NonGeneric": "4.0.1",
         "System.Console": "4.0.0",
         "System.Diagnostics.TraceSource": "4.0.0",