--- a/src/Tasks/project.json
+++ b/src/Tasks/project.json
@@ -26,6 +26,7 @@
         "System.Runtime.InteropServices.RuntimeInformation": "4.0.0",
         "System.Security.Cryptography.Algorithms": "4.2.0",
         "System.Text.Encoding.CodePages" : "4.0.1",
+        "System.Threading.Tasks.Parallel": "4.0.1",
         "System.Threading.Thread": "4.0.0",
         "System.Xml.XmlDocument": "4.0.1",
         "System.Xml.XPath": "4.0.1",