--- a/src/Utilities/project.json
+++ b/src/Utilities/project.json
@@ -5,12 +5,14 @@
         "Microsoft.VisualStudio.Setup.Configuration.Interop":  "1.14.114",
         "PdbGit": "3.0.41",
         "SourceLink.Create.CommandLine": "2.1.2",
+        "System.Collections.Immutable": "1.3.1",
         "System.Runtime.InteropServices.RuntimeInformation": "4.3.0"
       }
     },
     "netstandard1.3": {
       "dependencies": {
         "NETStandard.Library": "1.6.0",
+        "System.Collections.Immutable": "1.2.0",
         "System.Collections.NonGeneric": "4.0.1",
         "System.Diagnostics.Process": "4.1.0",
         "System.Diagnostics.TraceSource": "4.0.0",
@@ -20,6 +22,7 @@
         "System.Runtime.InteropServices.RuntimeInformation": "4.0.0",
         "System.Security.Cryptography.X509Certificates": "4.1.0",
         "System.Text.Encoding.CodePages" : "4.0.1",
+        "System.Threading.Tasks.Parallel": "4.0.1",
         "System.Threading.Thread": "4.0.0",
         "System.Xml.XmlDocument": "4.0.1"
       }