… src/Shared/FileMatcher.cs