@cdmihai 2017-10-03
Looking into the tests.