@maca88 2017-10-03
The problem is with the usage of `File.GetFileName` that doesn't treat a backslash as a director separator on Unix, so `Path.GetFileName("Folder\MyFile.cs")` will return `"Folder\MyFile.cs"` instead of `"MyFile.cs"`. My plan is to add a custom `GetFileName` that will work the same on Windows and Unix. Also, I suspect that `Path.GetDirectoryName` [here](https://github.com/Microsoft/msbuild/pull/2572/files#diff-64d47c5cbb59959b7b2fa6963672d596R827) will not work in all cases, I will do some investigation tomorrow.