@cdmihai 2017-10-03
I think the main issue here is the testing mock file system returns paths with backslashes in them, regardless of the OS. I'll try and make it normalize slashes.