@maca88 2017-10-04
Fixed directory matching and flattern the `AggregateException` in case the current settings for the parallel traversing will be altered to have nested `Parallel.ForEach` calls. With the current settings only one `Parallel.ForEach` is used per `GetFiles` call, as [MaxTasksPerIteration](https://github.com/Microsoft/msbuild/pull/2572/files#diff-64d47c5cbb59959b7b2fa6963672d596R2141) is set to the same value as `MaxTasks`.