@maca88 2017-10-05
There seems to be an inconsistent behavior with excludes and case sensibility. In the current master, when an exclude is matched via regex the matching is [case-insensitive](https://github.com/Microsoft/msbuild/blob/master/src/Shared/FileMatcher.cs#L1578) and when is matched with the [searchesToExcludeInSubdirs](https://github.com/Microsoft/msbuild/blob/master/src/Shared/FileMatcher.cs#L797) dictionary, is case-sensitive, as the dictionary [does not use](https://github.com/Microsoft/msbuild/blob/master/src/Shared/FileMatcher.cs#L1840) a case-insensitive comparer. For example, if we modify [this line](https://github.com/Microsoft/msbuild/blob/master/src/Shared/UnitTests/FileMatcher_Tests.cs#L1034) to `@"**\bIN\**\*.cs"` the test still succeeds, but if `@"src\Common\**"` is changed to `@"src\CoMMon\**"` the test fails. What should be the correct behavior? If case-insensitive is the answer, then I will modify the pattern matching algorithm to support it. Also, with this PR, an exclude like `@"src\Co??on\**"` will exclude files from `"src\Common"`, but in the in current master, it won't. Is this behavior desired, or should we change it in order to behave the same as in the current master?