@maca88 2017-10-06
@cdmihai 

> So to keep with previous behaviour, both matches need to be case insensitive.

I will modify the algorithm to support case-insensitive matching as fast as possible, correct the initialization of `searchesToExcludeInSubdirs` dictionary to use a case-insensitive comparer and add tests for complex patterns using the real file system.

> Regarding the double question mark match, master seems to work

My bad, I didn't noticed that the error was due to the mock file system, which doesn't support complex patterns.