@maca88 2017-10-06
@davkean

The exception handling should now work like before, with two exceptions:
1. If an exception is thrown inside a `Parallel.ForEach` and it is not handled by `ExceptionHandling.IsIoRelatedException`, then an `AggregateException` will be [thrown](https://github.com/maca88/msbuild/blob/b5574289426c99e9567a43fe17d4731e1f4f5380/src/Shared/FileMatcher.cs#L2227) instead of the actual exception/s.
2. By replacing the `Directory.GetXXX` with `Directory.EnumerateXXX` we may get one or more files/directories before the first `Security`/`UnauthorizedAccess` exception is thrown, in other words, if the third file is not accessible, with `Directory.EnumerateXXX` we would get the first two, but with `Directory.GetXXX` we wouldn't get any.