@cdmihai 2017-10-06
@dotnet-bot test Windows_NT Build for Full please

Flaky restore