@lifengl 2017-10-06
That is really nice work, @maca88.  I saw so many traces that the slow globbing code consumed 10 seconds or even 30 seconds in some projects.  This change will make a huge difference.