@maca88 2017-10-07
@cdmihai

I have corrected the case sensitivity behavior, so that everything is now matched case-insensitive. I've added a unit test for each correction related to matching excludes, which would fail on the current master. The overall performance after adding the case-insensitive matching remains more or less the same as before.